G2-666 Move bash model & renderer
