+++ en-backend/src/main/java/com/g2forge/enigma/backend/convert/common/IRenderer.java
@@ -1,5 +1,5 @@
 package com.g2forge.enigma.backend.convert.common;
 
-public interface IRenderer<R extends IRenderable> {
+public interface IRenderer<R> {
 	public String render(R renderable);
 }
