--- en-backend/src/test/java/com/g2forge/enigma/backend/bash/model/IBashBlock.java
@@ -1,5 +1,0 @@
-package com.g2forge.enigma.backend.bash.model;
-
-import com.g2forge.enigma.backend.bash.convert.IBashRenderable;
-
-public interface IBashBlock extends IBashRenderable {}
