--- en-backend/src/test/java/com/g2forge/enigma/backend/bash/model/expression/IBashExpression.java
@@ -1,3 +1,0 @@
-package com.g2forge.enigma.backend.bash.model.expression;
-
-public interface IBashExpression {}
