+++ en-bash/src/main/java/com/g2forge/enigma/bash/model/IBashBlock.java
@@ -1,0 +1,5 @@
+package com.g2forge.enigma.bash.model;
+
+import com.g2forge.enigma.bash.convert.IBashRenderable;
+
+public interface IBashBlock extends IBashRenderable {}
