+++ en-bash/src/main/java/com/g2forge/enigma/bash/model/IBashStatement.java
@@ -1,3 +1,3 @@
-package com.g2forge.enigma.backend.bash.model;
+package com.g2forge.enigma.bash.model;
 
 public interface IBashStatement extends IBashBlock {}
