+++ en-bash/src/main/java/com/g2forge/enigma/bash/model/expression/BashCommandSubstitution.java
@@ -1,6 +1,6 @@
-package com.g2forge.enigma.backend.bash.model.expression;
+package com.g2forge.enigma.bash.model.expression;
 
-import com.g2forge.enigma.backend.bash.model.BashCommand;
+import com.g2forge.enigma.bash.model.BashCommand;
 
 import lombok.Builder;
 import lombok.Data;
