+++ en-bash/src/main/java/com/g2forge/enigma/bash/model/expression/IBashExpression.java
@@ -1,0 +1,3 @@
+package com.g2forge.enigma.bash.model.expression;
+
+public interface IBashExpression {}
